+++ src/scorer/weight_presets.ts
@@ -1,0 +1,64 @@
+/*
+ * Preset configurations to simplify user weight management.
+ */
+
+import { WeightName, Weights } from "../types";
+
+
+export enum PresetWeightLabel {
+    DEFAULT = 'Default',
+    FRIENDS = 'Friends',
+    PICTURES = 'Pictures',
+    // FOLLOWING = 'Following',
+    TRENDING = 'Trending',
+};
+
+export type WeightPresets = Record<PresetWeightLabel, Weights>;
+
+export const DEFAULT_WEIGHTS: Weights = {
+    // Global modifiers that affect all weighted scores
+    [WeightName.TIME_DECAY]: 0.05,
+    [WeightName.TRENDING]: 0.06,
+    // Weighted scores
+    [WeightName.CHAOS]: 1.0,
+    [WeightName.DIVERSITY]: 1.0,
+    [WeightName.FAVORITED_ACCOUNTS]: 1.0,
+    [WeightName.FOLLOWED_TAGS]: 2.0,
+    [WeightName.IMAGE_ATTACHMENTS]: 0,
+    [WeightName.INTERACTIONS]: 1.5,
+    [WeightName.MENTIONS_FOLLOWED]: 2.0,
+    [WeightName.MOST_REPLIED_ACCOUNTS]: 1.0,
+    [WeightName.MOST_RETOOTED_ACCOUNTS]: 2,
+    [WeightName.NUM_FAVOURITES]: 1.0,
+    [WeightName.NUM_REPLIES]: 1.0,
+    [WeightName.NUM_RETOOTS]: 1.0,
+    [WeightName.RETOOTED_IN_FEED]: 2.0,
+    [WeightName.TRENDING_LINKS]: 0.7,
+    [WeightName.TRENDING_TAGS]: 0.5,
+    [WeightName.TRENDING_TOOTS]: 1.0,
+    [WeightName.VIDEO_ATTACHMENTS]: 0,
+};
+
+export const PresetWeights: WeightPresets = {
+    [PresetWeightLabel.DEFAULT]: DEFAULT_WEIGHTS,
+
+    [PresetWeightLabel.FRIENDS]: {
+        ...DEFAULT_WEIGHTS,
+        [WeightName.TRENDING]: 0.04,
+        [WeightName.INTERACTIONS]: 2.5,
+        [WeightName.MENTIONS_FOLLOWED]: 3.0,
+        [WeightName.MOST_REPLIED_ACCOUNTS]: 3.0,
+        [WeightName.MOST_RETOOTED_ACCOUNTS]: 3.0,
+    },
+
+    [PresetWeightLabel.PICTURES]: {
+        ...DEFAULT_WEIGHTS,
+        [WeightName.IMAGE_ATTACHMENTS]: 5.0,
+        [WeightName.VIDEO_ATTACHMENTS]: 5.0,
+    },
+
+    [PresetWeightLabel.TRENDING]: {
+        ...DEFAULT_WEIGHTS,
+        [WeightName.TRENDING]: DEFAULT_WEIGHTS[WeightName.TRENDING] * 3,
+    },
+};